+++ src/main/java/ar/edu/untref/gio/validator/TermDepositValidator.java
@@ -4,6 +4,6 @@ import java.util.Date;
 
 public interface TermDepositValidator {
 
-    void execute(Double amount, Double rate, Date expiration, Long creatorId);
+    void execute(Double amount, Double rate, Date expiration, Integer creatorId);
 
 }
